adds singleview and type list component, routes for type and product, productreducer.

+++ server/db/models/orderProduct.js
@@ -11,7 +11,7 @@ const OrderProduct = db.define('orderProduct', {
     }
   },
   price: {
-    type: Sequelize.DOUBLE,
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
       notEmpty: true
